changed file structure/added new folder for review sidebar
