+++ client/src/components/RatingsComponents/ReviewsList/ThumbnailModal.jsx
