+++ client/src/components/RatingsComponents/ReviewsSidebar/AverageStars.jsx
@@ -1,0 +1,1 @@
+import React from 'react';
